+++ src/views/select/select.jsx
@@ -1,17 +1,25 @@
-import {Button} from '../../components/index';
-import {CardTitle} from '../../components/styled';
-import {SelectionCard} from './styled';
+
+import {DiagramSelector} from './diagram-selector';
+import Card from '@mui/material/Card';
+import CardContent from '@mui/material/CardContent';
+import Typography from '@mui/material/Typography';
 
 const Select = (props) => {
 
     return (
-        <SelectionCard>
-            <CardTitle>Diagram Selector</CardTitle>
-            <p>Select the type of diagram you wish to create:</p>
-            <p>
-               <Button text="Flowchart" /> 
-            </p>
-        </SelectionCard>
+        <Card>
+
+            <CardContent>
+                <Typography variant="h5" component="div" gutterBottom>
+                    Select the type of Diagram to Build:
+                </Typography>
+
+                <DiagramSelector />
+
+            </CardContent>
+                        
+            
+        </Card>
     );
 };